Extract attributes actions to a library that can be shared (#1322)
