+++ processor/attributesprocessor/attributes_test.go
@@ -26,12 +26,13 @@ import (
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/exporter/exportertest"
 	"go.opentelemetry.io/collector/internal/data/testdata"
+	"go.opentelemetry.io/collector/internal/processor/attraction"
 	"go.opentelemetry.io/collector/internal/processor/filterset"
 	"go.opentelemetry.io/collector/internal/processor/filterspan"
 	"go.opentelemetry.io/collector/translator/conventions"
 )
 
-// Common structure for the
+// Common structure for all the Tests
 type testCase struct {
 	name               string
 	serviceName        string
@@ -137,9 +138,9 @@ func TestSpanProcessor_NilEmptyData(t *testing.T) {
 	factory := Factory{}
 	cfg := factory.CreateDefaultConfig()
 	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "attribute1", Action: INSERT, Value: 123},
-		{Key: "attribute1", Action: DELETE},
+	oCfg.Settings.Actions = []attraction.ActionKeyValue{
+		{Key: "attribute1", Action: attraction.INSERT, Value: 123},
+		{Key: "attribute1", Action: attraction.DELETE},
 	}
 
 	tp, err := factory.CreateTraceProcessor(
@@ -152,613 +153,6 @@ func TestSpanProcessor_NilEmptyData(t *testing.T) {
 			assert.NoError(t, tp.ConsumeTraces(context.Background(), tt.input))
 			assert.EqualValues(t, tt.output, tt.input)
 		})
-	}
-}
-
-func TestAttributes_InsertValue(t *testing.T) {
-	testCases := []testCase{
-		// Ensure `attribute1` is set for spans with no attributes.
-		{
-			name:            "InsertEmptyAttributes",
-			inputAttributes: map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"attribute1": pdata.NewAttributeValueInt(123),
-			},
-		},
-		// Ensure `attribute1` is set.
-		{
-			name: "InsertKeyNoExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"anotherkey": pdata.NewAttributeValueString("bob"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"anotherkey": pdata.NewAttributeValueString("bob"),
-				"attribute1": pdata.NewAttributeValueInt(123),
-			},
-		},
-		// Ensures no insert is performed because the keys `attribute1` already exists.
-		{
-			name: "InsertKeyExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"attribute1": pdata.NewAttributeValueString("bob"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"attribute1": pdata.NewAttributeValueString("bob"),
-			},
-		},
-	}
-
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "attribute1", Action: INSERT, Value: 123},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_InsertFromAttribute(t *testing.T) {
-
-	testCases := []testCase{
-		// Ensure no attribute is inserted because because attributes do not exist.
-		{
-			name:               "InsertEmptyAttributes",
-			inputAttributes:    map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{},
-		},
-		// Ensure no attribute is inserted because because from_attribute `string_key` does not exist.
-		{
-			name: "InsertMissingFromAttribute",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"bob": pdata.NewAttributeValueInt(1),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"bob": pdata.NewAttributeValueInt(1),
-			},
-		},
-		// Ensure `string key` is set.
-		{
-			name: "InsertAttributeExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"anotherkey": pdata.NewAttributeValueInt(8892342),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"anotherkey": pdata.NewAttributeValueInt(8892342),
-				"string key": pdata.NewAttributeValueInt(8892342),
-			},
-		},
-		// Ensures no insert is performed because the keys `string key` already exist.
-		{
-			name: "InsertKeysExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"anotherkey": pdata.NewAttributeValueInt(8892342),
-				"string key": pdata.NewAttributeValueString("here"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"anotherkey": pdata.NewAttributeValueInt(8892342),
-				"string key": pdata.NewAttributeValueString("here"),
-			},
-		},
-	}
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "string key", Action: INSERT, FromAttribute: "anotherkey"},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_UpdateValue(t *testing.T) {
-
-	testCases := []testCase{
-		// Ensure no changes to the span as there is no attributes map.
-		{
-			name:               "UpdateNoAttributes",
-			inputAttributes:    map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{},
-		},
-		// Ensure no changes to the span as the key does not exist.
-		{
-			name: "UpdateKeyNoExist",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("foo"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("foo"),
-			},
-		},
-		// Ensure the attribute `db.secret` is updated.
-		{
-			name: "UpdateAttributes",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"db.secret": pdata.NewAttributeValueString("password1234"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"db.secret": pdata.NewAttributeValueString("redacted"),
-			},
-		},
-	}
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "db.secret", Action: UPDATE, Value: "redacted"},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_UpdateFromAttribute(t *testing.T) {
-
-	testCases := []testCase{
-		// Ensure no changes to the span as there is no attributes map.
-		{
-			name:               "UpdateNoAttributes",
-			inputAttributes:    map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{},
-		},
-		// Ensure the attribute `boo` isn't updated because attribute `foo` isn't present in the span.
-		{
-			name: "UpdateKeyNoExistFromAttribute",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("bob"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("bob"),
-			},
-		},
-		// Ensure no updates as the target key `boo` doesn't exists.
-		{
-			name: "UpdateKeyNoExistMainAttributed",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"foo": pdata.NewAttributeValueString("over there"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"foo": pdata.NewAttributeValueString("over there"),
-			},
-		},
-		// Ensure no updates as the target key `boo` doesn't exists.
-		{
-			name: "UpdateKeyFromExistingAttribute",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"foo": pdata.NewAttributeValueString("there is a party over here"),
-				"boo": pdata.NewAttributeValueString("not here"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"foo": pdata.NewAttributeValueString("there is a party over here"),
-				"boo": pdata.NewAttributeValueString("there is a party over here"),
-			},
-		},
-	}
-
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "boo", Action: UPDATE, FromAttribute: "foo"},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_UpsertValue(t *testing.T) {
-	testCases := []testCase{
-		// Ensure `region` is set for spans with no attributes.
-		{
-			name:            "UpsertNoAttributes",
-			inputAttributes: map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"region": pdata.NewAttributeValueString("planet-earth"),
-			},
-		},
-		// Ensure `region` is inserted for spans with some attributes(the key doesn't exist).
-		{
-			name: "UpsertAttributeNoExist",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"mission": pdata.NewAttributeValueString("to mars"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"mission": pdata.NewAttributeValueString("to mars"),
-				"region":  pdata.NewAttributeValueString("planet-earth"),
-			},
-		},
-		/// Ensure `region` is updated for spans with the attribute key `region`.
-		{
-			name: "UpsertAttributeExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"mission": pdata.NewAttributeValueString("to mars"),
-				"region":  pdata.NewAttributeValueString("solar system"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"mission": pdata.NewAttributeValueString("to mars"),
-				"region":  pdata.NewAttributeValueString("planet-earth"),
-			},
-		},
-	}
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "region", Action: UPSERT, Value: "planet-earth"},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_Extract(t *testing.T) {
-	testCases := []testCase{
-		// Ensure `new_user_key` is not set for spans with no attributes.
-		{
-			name:               "UpsertEmptyAttributes",
-			inputAttributes:    map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{},
-		},
-		// Ensure `new_user_key` is not inserted for spans with missing attribute `user_key`.
-		{
-			name: "No extract with no target key",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("ghosts are scary"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("ghosts are scary"),
-			},
-		},
-		// Ensure `new_user_key` is not inserted for spans with missing attribute `user_key`.
-		{
-			name: "No extract with non string target key",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"boo":      pdata.NewAttributeValueString("ghosts are scary"),
-				"user_key": pdata.NewAttributeValueInt(1234),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"boo":      pdata.NewAttributeValueString("ghosts are scary"),
-				"user_key": pdata.NewAttributeValueInt(1234),
-			},
-		},
-		// Ensure `new_user_key` is not updated for spans with attribute
-		// `user_key` because `user_key` does not match the regular expression.
-		{
-			name: "No extract with no pattern matching",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key": pdata.NewAttributeValueString("does not match"),
-				"boo":      pdata.NewAttributeValueString("ghosts are scary"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key": pdata.NewAttributeValueString("does not match"),
-				"boo":      pdata.NewAttributeValueString("ghosts are scary"),
-			},
-		},
-		// Ensure `new_user_key` is not updated for spans with attribute
-		// `user_key` because `user_key` does not match all of the regular
-		// expression.
-		{
-			name: "No extract with no pattern matching",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key": pdata.NewAttributeValueString("/api/v1/document/12345678/update"),
-				"boo":      pdata.NewAttributeValueString("ghosts are scary"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key": pdata.NewAttributeValueString("/api/v1/document/12345678/update"),
-				"boo":      pdata.NewAttributeValueString("ghosts are scary"),
-			},
-		},
-		// Ensure `new_user_key` and `version` is inserted for spans with attribute `user_key`.
-		{
-			name: "Extract insert new values.",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key": pdata.NewAttributeValueString("/api/v1/document/12345678/update/v1"),
-				"foo":      pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueString("/api/v1/document/12345678/update/v1"),
-				"new_user_key": pdata.NewAttributeValueString("12345678"),
-				"version":      pdata.NewAttributeValueString("v1"),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-		},
-		// Ensure `new_user_key` and `version` is updated for spans with attribute `user_key`.
-		{
-			name: "Extract updates existing values ",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueString("/api/v1/document/12345678/update/v1"),
-				"new_user_key": pdata.NewAttributeValueString("2321"),
-				"version":      pdata.NewAttributeValueString("na"),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueString("/api/v1/document/12345678/update/v1"),
-				"new_user_key": pdata.NewAttributeValueString("12345678"),
-				"version":      pdata.NewAttributeValueString("v1"),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-		},
-		// Ensure `new_user_key` is updated and `version` is inserted for spans with attribute `user_key`.
-		{
-			name: "Extract upserts values",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueString("/api/v1/document/12345678/update/v1"),
-				"new_user_key": pdata.NewAttributeValueString("2321"),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueString("/api/v1/document/12345678/update/v1"),
-				"new_user_key": pdata.NewAttributeValueString("12345678"),
-				"version":      pdata.NewAttributeValueString("v1"),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-		},
-	}
-
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "user_key", RegexPattern: "^\\/api\\/v1\\/document\\/(?P<new_user_key>.*)\\/update\\/(?P<version>.*)$", Action: EXTRACT},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_UpsertFromAttribute(t *testing.T) {
-
-	testCases := []testCase{
-		// Ensure `new_user_key` is not set for spans with no attributes.
-		{
-			name:               "UpsertEmptyAttributes",
-			inputAttributes:    map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{},
-		},
-		// Ensure `new_user_key` is not inserted for spans with missing attribute `user_key`.
-		{
-			name: "UpsertFromAttributeNoExist",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("ghosts are scary"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("ghosts are scary"),
-			},
-		},
-		// Ensure `new_user_key` is inserted for spans with attribute `user_key`.
-		{
-			name: "UpsertFromAttributeExistsInsert",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key": pdata.NewAttributeValueInt(2245),
-				"foo":      pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueInt(2245),
-				"new_user_key": pdata.NewAttributeValueInt(2245),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-		},
-		// Ensure `new_user_key` is updated for spans with attribute `user_key`.
-		{
-			name: "UpsertFromAttributeExistsUpdate",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueInt(2245),
-				"new_user_key": pdata.NewAttributeValueInt(5422),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"user_key":     pdata.NewAttributeValueInt(2245),
-				"new_user_key": pdata.NewAttributeValueInt(2245),
-				"foo":          pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-		},
-	}
-
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "new_user_key", Action: UPSERT, FromAttribute: "user_key"},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_Delete(t *testing.T) {
-	testCases := []testCase{
-		// Ensure the span contains no changes.
-		{
-			name:               "DeleteEmptyAttributes",
-			inputAttributes:    map[string]pdata.AttributeValue{},
-			expectedAttributes: map[string]pdata.AttributeValue{},
-		},
-		// Ensure the span contains no changes because the key doesn't exist.
-		{
-			name: "DeleteAttributeNoExist",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("ghosts are scary"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"boo": pdata.NewAttributeValueString("ghosts are scary"),
-			},
-		},
-		// Ensure `duplicate_key` is deleted for spans with the attribute set.
-		{
-			name: "DeleteAttributeExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"duplicate_key": pdata.NewAttributeValueDouble(3245.6),
-				"original_key":  pdata.NewAttributeValueDouble(3245.6),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"original_key": pdata.NewAttributeValueDouble(3245.6),
-			},
-		},
-	}
-
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "duplicate_key", Action: DELETE},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
-	}
-}
-
-func TestAttributes_FromAttributeNoChange(t *testing.T) {
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "boo", Action: INSERT, FromAttribute: "boo"},
-		{Key: "boo", Action: UPDATE, FromAttribute: "boo"},
-		{Key: "boo", Action: UPSERT, FromAttribute: "boo"},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-	traceData := generateTraceData(
-		"",
-		"FromAttributeNoChange",
-		map[string]pdata.AttributeValue{
-			"boo": pdata.NewAttributeValueString("ghosts are scary"),
-		})
-
-	assert.NoError(t, tp.ConsumeTraces(context.Background(), traceData))
-	require.Equal(t, generateTraceData(
-		"",
-		"FromAttributeNoChange",
-		map[string]pdata.AttributeValue{
-			"boo": pdata.NewAttributeValueString("ghosts are scary"),
-		}), traceData)
-}
-
-func TestAttributes_Ordering(t *testing.T) {
-	testCases := []testCase{
-		// For this example, the operations performed are
-		// 1. insert `operation`: `default`
-		// 2. insert `svc.operation`: `default`
-		// 3. delete `operation`.
-		{
-			name: "OrderingApplyAllSteps",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"foo": pdata.NewAttributeValueString("casper the friendly ghost"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"foo":           pdata.NewAttributeValueString("casper the friendly ghost"),
-				"svc.operation": pdata.NewAttributeValueString("default"),
-			},
-		},
-		// For this example, the operations performed are
-		// 1. do nothing for the first action of insert `operation`: `default`
-		// 2. insert `svc.operation`: `arithmetic`
-		// 3. delete `operation`.
-		{
-			name: "OrderingOperationExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"foo":       pdata.NewAttributeValueString("casper the friendly ghost"),
-				"operation": pdata.NewAttributeValueString("arithmetic"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"foo":           pdata.NewAttributeValueString("casper the friendly ghost"),
-				"svc.operation": pdata.NewAttributeValueString("arithmetic"),
-			},
-		},
-
-		// For this example, the operations performed are
-		// 1. insert `operation`: `default`
-		// 2. update `svc.operation` to `default`
-		// 3. delete `operation`.
-		{
-			name: "OrderingSvcOperationExists",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"foo":           pdata.NewAttributeValueString("casper the friendly ghost"),
-				"svc.operation": pdata.NewAttributeValueString("some value"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"foo":           pdata.NewAttributeValueString("casper the friendly ghost"),
-				"svc.operation": pdata.NewAttributeValueString("default"),
-			},
-		},
-
-		// For this example, the operations performed are
-		// 1. do nothing for the first action of insert `operation`: `default`
-		// 2. update `svc.operation` to `arithmetic`
-		// 3. delete `operation`.
-		{
-			name: "OrderingBothAttributesExist",
-			inputAttributes: map[string]pdata.AttributeValue{
-				"foo":           pdata.NewAttributeValueString("casper the friendly ghost"),
-				"operation":     pdata.NewAttributeValueString("arithmetic"),
-				"svc.operation": pdata.NewAttributeValueString("add"),
-			},
-			expectedAttributes: map[string]pdata.AttributeValue{
-				"foo":           pdata.NewAttributeValueString("casper the friendly ghost"),
-				"svc.operation": pdata.NewAttributeValueString("arithmetic"),
-			},
-		},
-	}
-
-	factory := Factory{}
-	cfg := factory.CreateDefaultConfig()
-	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "operation", Action: INSERT, Value: "default"},
-		{Key: "svc.operation", Action: UPSERT, FromAttribute: "operation"},
-		{Key: "operation", Action: DELETE},
-	}
-
-	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
-	require.Nil(t, err)
-	require.NotNil(t, tp)
-
-	for _, tt := range testCases {
-		runIndividualTestCase(t, tt, tp)
 	}
 }
 
@@ -804,8 +198,8 @@ func TestAttributes_FilterSpans(t *testing.T) {
 	factory := Factory{}
 	cfg := factory.CreateDefaultConfig()
 	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "attribute1", Action: INSERT, Value: 123},
+	oCfg.Actions = []attraction.ActionKeyValue{
+		{Key: "attribute1", Action: attraction.INSERT, Value: 123},
 	}
 	oCfg.Include = &filterspan.MatchProperties{
 		Services: []string{"svcA", "svcB.*"},
@@ -874,8 +268,8 @@ func TestAttributes_FilterSpansByNameStrict(t *testing.T) {
 	factory := Factory{}
 	cfg := factory.CreateDefaultConfig()
 	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "attribute1", Action: INSERT, Value: 123},
+	oCfg.Actions = []attraction.ActionKeyValue{
+		{Key: "attribute1", Action: attraction.INSERT, Value: 123},
 	}
 	oCfg.Include = &filterspan.MatchProperties{
 		SpanNames: []string{"apply", "dont_apply"},
@@ -942,8 +336,8 @@ func TestAttributes_FilterSpansByNameRegexp(t *testing.T) {
 	factory := Factory{}
 	cfg := factory.CreateDefaultConfig()
 	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "attribute1", Action: INSERT, Value: 123},
+	oCfg.Actions = []attraction.ActionKeyValue{
+		{Key: "attribute1", Action: attraction.INSERT, Value: 123},
 	}
 	oCfg.Include = &filterspan.MatchProperties{
 		SpanNames: []string{"^apply.*"},
@@ -1005,11 +399,11 @@ func TestAttributes_Hash(t *testing.T) {
 	factory := Factory{}
 	cfg := factory.CreateDefaultConfig()
 	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "user.email", Action: HASH},
-		{Key: "user.id", Action: HASH},
-		{Key: "user.balance", Action: HASH},
-		{Key: "user.authenticated", Action: HASH},
+	oCfg.Actions = []attraction.ActionKeyValue{
+		{Key: "user.email", Action: attraction.HASH},
+		{Key: "user.id", Action: attraction.HASH},
+		{Key: "user.balance", Action: attraction.HASH},
+		{Key: "user.authenticated", Action: attraction.HASH},
 	}
 
 	tp, err := factory.CreateTraceProcessor(context.Background(), component.ProcessorCreateParams{}, exportertest.NewNopTraceExporter(), cfg)
@@ -1050,8 +444,8 @@ func BenchmarkAttributes_FilterSpansByName(b *testing.B) {
 	factory := Factory{}
 	cfg := factory.CreateDefaultConfig()
 	oCfg := cfg.(*Config)
-	oCfg.Actions = []ActionKeyValue{
-		{Key: "attribute1", Action: INSERT, Value: 123},
+	oCfg.Actions = []attraction.ActionKeyValue{
+		{Key: "attribute1", Action: attraction.INSERT, Value: 123},
 	}
 	oCfg.Include = &filterspan.MatchProperties{
 		SpanNames: []string{"^apply.*"},
