+++ processor/attributesprocessor/factory.go
@@ -17,14 +17,12 @@ package attributesprocessor
 import (
 	"context"
 	"fmt"
-	"regexp"
-	"strings"
 
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/configerror"
 	"go.opentelemetry.io/collector/config/configmodels"
 	"go.opentelemetry.io/collector/consumer"
-	"go.opentelemetry.io/collector/internal/processor/filterhelper"
+	"go.opentelemetry.io/collector/internal/processor/attraction"
 	"go.opentelemetry.io/collector/internal/processor/filterspan"
 )
 
@@ -62,9 +60,12 @@ func (f *Factory) CreateTraceProcessor(
 ) (component.TraceProcessor, error) {
 
 	oCfg := cfg.(*Config)
-	actions, err := buildAttributesConfiguration(*oCfg)
+	if len(oCfg.Actions) == 0 {
+		return nil, fmt.Errorf("error creating \"attributes\" processor due to missing required field \"actions\" of processor %q", cfg.Name())
+	}
+	attrProc, err := attraction.NewAttrProc(&oCfg.Settings)
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("error creating \"attributes\" processor: %w of processor %q", err, cfg.Name())
 	}
 	include, err := filterspan.NewMatcher(oCfg.Include)
 	if err != nil {
@@ -74,12 +75,7 @@ func (f *Factory) CreateTraceProcessor(
 	if err != nil {
 		return nil, err
 	}
-	config := attributesConfig{
-		actions: actions,
-		include: include,
-		exclude: exclude,
-	}
-	return newTraceProcessor(nextConsumer, config)
+	return newTraceProcessor(nextConsumer, attrProc, include, exclude)
 }
 
 // CreateMetricsProcessor creates a metrics processor based on this config.
@@ -90,86 +86,4 @@ func (f *Factory) CreateMetricsProcessor(
 	_ configmodels.Processor,
 ) (component.MetricsProcessor, error) {
 	return nil, configerror.ErrDataTypeIsNotSupported
-}
-
-// buildAttributesConfiguration validates the input configuration has all of the required fields for the processor
-// and returns a list of valid actions to configure the processor.
-// An error is returned if there are any invalid inputs.
-func buildAttributesConfiguration(config Config) ([]attributeAction, error) {
-	if len(config.Actions) == 0 {
-		return nil, fmt.Errorf("error creating \"attributes\" processor due to missing required field \"actions\" of processor %q", config.Name())
-	}
-
-	var attributeActions []attributeAction
-	for i, a := range config.Actions {
-		// `key` is a required field
-		if a.Key == "" {
-			return nil, fmt.Errorf("error creating \"attributes\" processor due to missing required field \"key\" at the %d-th actions of processor %q", i, config.Name())
-		}
-
-		// Convert `action` to lowercase for comparison.
-		a.Action = Action(strings.ToLower(string(a.Action)))
-		action := attributeAction{
-			Key:    a.Key,
-			Action: a.Action,
-		}
-
-		switch a.Action {
-		case INSERT, UPDATE, UPSERT:
-			if a.Value == nil && a.FromAttribute == "" {
-				return nil, fmt.Errorf("error creating \"attributes\" processor. Either field \"value\" or \"from_attribute\" setting must be specified for %d-th action of processor %q", i, config.Name())
-			}
-
-			if a.Value != nil && a.FromAttribute != "" {
-				return nil, fmt.Errorf("error creating \"attributes\" processor due to both fields \"value\" and \"from_attribute\" being set at the %d-th actions of processor %q", i, config.Name())
-			}
-			if a.RegexPattern != "" {
-				return nil, fmt.Errorf("error creating \"attributes\" processor. Action \"%s\" does not use the \"pattern\" field. This must not be specified for %d-th action of processor %q", a.Action, i, config.Name())
-
-			}
-			// Convert the raw value from the configuration to the internal trace representation of the value.
-			if a.Value != nil {
-				val, err := filterhelper.NewAttributeValueRaw(a.Value)
-				if err != nil {
-					return nil, err
-				}
-				action.AttributeValue = &val
-			} else {
-				action.FromAttribute = a.FromAttribute
-			}
-		case HASH, DELETE:
-			if a.Value != nil || a.FromAttribute != "" || a.RegexPattern != "" {
-				return nil, fmt.Errorf("error creating \"attributes\" processor. Action \"%s\" does not use \"value\", \"pattern\" or \"from_attribute\" field. These must not be specified for %d-th action of processor %q", a.Action, i, config.Name())
-			}
-		case EXTRACT:
-			if a.Value != nil || a.FromAttribute != "" {
-				return nil, fmt.Errorf("error creating \"attributes\" processor. Action \"%s\" does not use \"value\" or \"from_attribute\" field. These must not be specified for %d-th action of processor %q", a.Action, i, config.Name())
-			}
-			if a.RegexPattern == "" {
-				return nil, fmt.Errorf("error creating \"attributes\" processor due to missing required field \"pattern\" for action \"%s\" at the %d-th action of processor %q", a.Action, i, config.Name())
-
-			}
-			re, err := regexp.Compile(a.RegexPattern)
-			if err != nil {
-				return nil, fmt.Errorf("error creating \"attributes\" processor. Field \"pattern\" has invalid pattern: \"%s\" to be set at the %d-th actions of processor %q", a.RegexPattern, i, config.Name())
-			}
-			attrNames := re.SubexpNames()
-			if len(attrNames) <= 1 {
-				return nil, fmt.Errorf("error creating \"attributes\" processor. Field \"pattern\" contains no named matcher groups at the %d-th actions of processor %q", i, config.Name())
-			}
-
-			for subExpIndex := 1; subExpIndex < len(attrNames); subExpIndex++ {
-				if attrNames[subExpIndex] == "" {
-					return nil, fmt.Errorf("error creating \"attributes\" processor. Field \"pattern\" contains at least one unnamed matcher group at the %d-th actions of processor %q", i, config.Name())
-				}
-			}
-			action.Regex = re
-			action.AttrNames = attrNames
-		default:
-			return nil, fmt.Errorf("error creating \"attributes\" processor due to unsupported action %q at the %d-th actions of processor %q", a.Action, i, config.Name())
-		}
-
-		attributeActions = append(attributeActions, action)
-	}
-	return attributeActions, nil
 }
